+++ src/_struct_param.jl
@@ -114,9 +114,7 @@ end
 
 Base.@kwdef mutable struct ParCovsel
     nlv::Int = 1    
-    meth::Symbol = :cov  
-    scalx::Bool = false
-    scaly::Bool = false
+    scal::Bool = false
 end 
 
 Base.@kwdef mutable struct ParRp